moved TxnCoordinator to its own mod
